fix: first connection with DB

+++ src/main/java/org/cheeseapp/repos/ClientRepo.java
@@ -3,5 +3,5 @@ package org.cheeseapp.repos;
 import org.cheeseapp.domain.Client;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ClientRepo extends CrudRepository<Client, Long> {
+public interface ClientRepo extends CrudRepository<Client, Integer> {
 }
